backend/requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
boto3
python-multipart
numpy
